requirements.txt
click==8.1.7
gradio==3.50.2
markdown-it-py==3.0.0
mdurl==0.1.2
psycopg2-binary==2.9.9
Pygments==2.16.1
rich==13.6.0
typer==0.9.0
typing_extensions==4.8.0
